% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterization.R
\name{rasterizeEachClusterVer}
\alias{rasterizeEachClusterVer}
\title{Rasterize each cluster and return the integrated rasters (mosaic)
when the plot is veritical setting.}
\usage{
rasterizeEachClusterVer(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of the extent.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/slide120_D1")
mosaicIntegration <- rasterizeEachClusterVer(sample)
}
