% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrep.R
\name{rasterPrep}
\alias{rasterPrep}
\title{Finding the extent with buffer applied.}
\usage{
rasterPrep(sample, n)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}

\item{n}{buffer for extent}
}
\value{
raster layer with calculated resolution and extent with buffer applied
}
\description{
Finding the extent with buffer applied.
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
raster <- rasterPrep(sample, 10)
}
