% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrep.R
\name{rasterPrepVer}
\alias{rasterPrepVer}
\title{Finding the size of extent when the coordinates are veritical setting}
\usage{
rasterPrepVer(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of the extent.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/slide120_D1")
raster <- rasterPrepVer(sample)
}
