% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolutionCalc}
\alias{resolutionCalc}
\title{Automatically calculating the resolution for rasterization.}
\usage{
resolutionCalc(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of resolution.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
List <- resolutionCalc(sample)
}
