% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolutionCalcVer}
\alias{resolutionCalcVer}
\title{Automatically calculating the resolution for rasterization
when the coordinates are veritical setting}
\usage{
resolutionCalcVer(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of resolution.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/slide120_D1")
List <- resolutionCalcVer(sample)
}
