% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataPrep.R, R/test.R
\name{spatialDataPrep}
\alias{spatialDataPrep}
\title{Prepare input data}
\usage{
spatialDataPrep(filename)

spatialDataPrep(filename)
}
\arguments{
\item{filename}{folder pathway that contains feature-bc-mstrix, analysis and spatial information.}
}
\value{
seruat  object that have cluster labels attached.

seruat  object that have cluster labels attached.
}
\description{
This function ---

This function ---
}
\examples{
filename <- "/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer"
sample <- spatialDataPrep(filename)
filename <- "/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer"
sample <- spatialDataPrep(filename)
}
