% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreHeatmap.R
\name{zscoreMatrix}
\alias{zscoreMatrix}
\title{Z-score matrix.}
\usage{
zscoreMatrix(sample, joincount.result)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached}

\item{joincount.result}{calculated result from join count analysis}
}
\value{
z-score matrix
}
\description{
Z-score matrix.
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
mosaicIntegration <- rasterizeEachCluster(sample)
joincount.result <- joincountAnalysis(mosaicIntegration)
matrix <- zscoreMatrix(sample, joincount.result)
}
