% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount.R
\name{joincountAnalysis}
\alias{joincountAnalysis}
\title{Join count analysis.}
\usage{
joincountAnalysis(mosaicIntegration)
}
\arguments{
\item{mosaicIntegration}{RasterLayer that output from Function rasterization.}
}
\value{
Join count analysis results - cluster pair vs. their z-score.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
mosaicIntegration <- rasterizeEachCluster(sample)
joincount.filtered <- joincountAnalysis(mosaicIntegration)
}
