% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mosaicIntPlot}
\alias{mosaicIntPlot}
\title{Visulization of the rasterization results}
\usage{
mosaicIntPlot(mosaicIntegration)
}
\arguments{
\item{mosaicIntegration}{Integrated rasters of each cluster.}
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
mosaicIntegration <- rasterizeEachCluster(sample)
mosaicIntPlot(mosaicIntegration)

}
