% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterization.R
\name{rasterizeEachCluster}
\alias{rasterizeEachCluster}
\title{Rasterize each cluster and return the integrated mosaic}
\usage{
rasterizeEachCluster(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
Integrated mosaic.
}
\description{
Rasterize each cluster and return the integrated mosaic
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
mosaicIntegration <- rasterizeEachCluster(sample)
}
