% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrep.R
\name{rasterPrep}
\alias{rasterPrep}
\title{Finding the size of extent}
\usage{
rasterPrep(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of the extent.
}
\description{
This function ---
}
\examples{
sample <- spatialDataPrep("/Users/ninasong/Desktop/Craig_lab/GeoSpatial/breast_cancer")
raster <- rasterPrep(sample)
}
